student_management/requirements.txt
# Django==3.2 
# djangorestframework 
# djongo 
Django==3.2
djangorestframework==3.12.4
djongo==1.3.6
pymongo==3.11.4

